refactor(cpu): migrated to 64bit word length, added new instructions
- migrated every part to 64bit word size
- made address space 32bit
- added new instructions
- improved a lot of loops

+++ src/main/java/de/happybavarian07/computer/assembler/encoder/model/EncodedWord.java
@@ -4,7 +4,7 @@ package de.happybavarian07.computer.assembler.encoder.model;
  * @Author HappyBavarian07
  * @Date August 11, 2026 | 20:11
  */
-public record EncodedWord(int byteAddress, int rawWord) implements Comparable<EncodedWord> {
+public record EncodedWord(int byteAddress, long rawWord) implements Comparable<EncodedWord> {
     @Override
     public int compareTo(EncodedWord other) {
         return Integer.compare(byteAddress, other.byteAddress);
